restore exhaustive union check, use @flow strict, fix degenSum deps
Exhaustive unions used to be commented out. The functionality has been restored,
and a test has been added to ensure its functionality. This is the first test in
the library. Now that we've laid the groundwork for it, we can expect more tests
in the future.

Both the source of flow-degen itself and the generated files now use the @flow
strict annotation. If your project doesn't use @flow strict, this should still
work since we strive to achieve maximum strictness with the generated files.

degenSum was not chaining its dependencies from degenEnum. To work around this,
consumers could add deString to their imports. This should be fixed now, but is
largely untested.

As part of the changes for testing, codeGen has been broken out. The codeGen
function now returns the generated code. Use the new fileGen function to
generate _and_ write to a file. This allows us to test our generated results
without hitting disk and causing Flow to identify errors at develop-time that we
intentionally create as part of running a test. We run Flow with check-contents,
which accepts a file from stdin.