Improve Flow Strict Compatibility
When the flow-strict options untyped-import or untyped-type-import are
enabled, flow doesn't seem to recognize .flow files that aren't in the
same directory as the corresponding .js file. This commit moves the
index.js.flow file into the dist folder alongside its index.js. As I
turned on the recommended set of strict options, a few other files
needed to be copied to dist/ as well and pulled in the libdef for
@babel/register.