fix strict fix
The strict fix applied earlier must have some other involved component, because
we've encountered projects that both use and don't use strict that the strict
fix seems to break. It is not well understood why or how switching to strict
mode breaks in some cases, and it is also not understood what those exact cases
are, but is not enough to simply be strict. This change brings back the original
index.js.flow workaround that lives in the root directory, and also retains the
original strict fix (because that fixed _something_).

Once we have more understanding we should record that somewhere.

Copious amounts of comments have been added to speak of what fixes are used
where and why. The Flow ticket that includes the index.js.flow workaround has
also been linked as well, per the original TODO.