Fix Error Running Flow Degen
The type annotations added to make sure that the call to fileGen in
flow-degen.js is typed correctly break when running flow-degen as plain
node does like the `import` or type annotation. The flow code is now
wrapped in flow comment blocks so that node will ignore it. This seemed
simpler than changing flow-degen.js to be transpiled by babel as part of
flow-dist.sh.