use Ramda mergeRight and fix Flow recursion issues
Ramda's merge no longer exists. After switching to mergeRight, I ran into
recursion limit issues. This is unlikely to be mergeRight's doing, as it is the
same definition. That said, the Ramda libdef might be more complicated, or
perhaps the latest Flow is less tolerant of recursion in the definitions.

Additionally, we shouldn't have to set the babelrcRoots anymore in part because
we're not using .babelrc anymore. The .babelrc file has different behavior from
babel.config.js relating to how they are employed across subdirectories and
nesting. I haven't confirmed this, and have instead leaned on the type
definition for the Babel config (Options) to define this.