use the prepare script target
This was unbeknownst to me at the time as a package maintainer:
"prepare" is a special chance for a package to produce any artifacts
require as part of its build. This part I knew. What I didn't know was
that prepare is executed as a special case when the package being
consumed is consumed over a git repository on disk. In the case of a
distributed package over something such as the npm repository, the
prepare statement is run at time of running npm/yarn publish. In the
case of a package consumed over git or on disk this means the prepare
script is run automatically immediately after the package has been
cloned from the git repository.

So instead of doing some of the machinery ourselves, this change puts us
on prepare. Since the build is running on a target machine and no longer
on Logan's personal machine, I've also converted flow-dist.sh to be a JS
file which should be platform agnostic.

The build artifacts are ignored now because they are produced in all
consumption scenarios, which has been a common source of confusion for
new contributors.