add github action release
Addresses #40 - add Github Action for handling releases.

This adds a workflow for tagging and publishing new packages for
flow-degen using Github Actions. This is done in hopes of automating
some of the publishing, which should occur far more frequently than it
has due to all of the dependabot security updates.